+++ core/src/commonMain/kotlin/com/boswelja/ephemeris/core/data/CalendarPageSource.kt
@@ -1,14 +1,14 @@
 package com.boswelja.ephemeris.core.data
 
-import com.boswelja.ephemeris.core.chunked
-import com.boswelja.ephemeris.core.endOfWeek
-import com.boswelja.ephemeris.core.map
+import com.boswelja.ephemeris.core.datetime.chunked
+import com.boswelja.ephemeris.core.datetime.endOfWeek
+import com.boswelja.ephemeris.core.datetime.map
 import com.boswelja.ephemeris.core.model.DisplayDate
-import com.boswelja.ephemeris.core.model.YearMonth
-import com.boswelja.ephemeris.core.model.plus
-import com.boswelja.ephemeris.core.model.until
-import com.boswelja.ephemeris.core.model.yearMonth
-import com.boswelja.ephemeris.core.startOfWeek
+import com.boswelja.ephemeris.core.datetime.YearMonth
+import com.boswelja.ephemeris.core.datetime.plus
+import com.boswelja.ephemeris.core.datetime.until
+import com.boswelja.ephemeris.core.datetime.yearMonth
+import com.boswelja.ephemeris.core.datetime.startOfWeek
 import kotlinx.datetime.Clock
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.DayOfWeek
