+++ core/src/commonMain/kotlin/com/boswelja/ephemeris/core/datetime/LocalDateExt.kt
@@ -1,4 +1,4 @@
-package com.boswelja.ephemeris.core
+package com.boswelja.ephemeris.core.datetime
 
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.DayOfWeek
